+++ examples/iframe.js
@@ -7,12 +7,14 @@ chromy.chain()
       .console(msg => console.log(msg))
       .iframe('iframe', async iframe => {
         await iframe.chain()
+                    .type('#txt', 'abc')
                     .click('#btn')
                     .evaluate(_ => {
                       console.log(document.querySelector('h1').innerText)
                     })
                     .end()
       })
+      .sleep(3000)
       .end()
       .catch(e => console.log(e))
       .then(_ => {